backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1